Update SteamVR_Teleporter.cs
added 

hasGroundTarget= true;

to line 71.  There are probably better ways to determine if you want to proceed, but this is better than the alternative where it will never work.

Altered the calculation for the new Camera Rig position on line 84.  Believe it should be contact world point of raycast offset by local point of hmd.  This should move the player to the targeted point as opposed to the current implementation where it appears to move Camera rig to targeted point offset by hmd and controller, which is disorienting.